improved css, refactored for cleanliness, implemented more answered qs feature
